Sound system
Added in a soundevent script and soundListener script the event object and listener.

Lister needs to have a rigidbody and the object creating the sound needs to have a sphere collider.

The values for the sound can be modified in the editor as they are public vars and DO NOT have default values in start.