Update Tony and Vision
Reorganized comments to be structured better.  Added a boolean that checks if Tony is idle or not. Moved Start() goals to after the goals list is actually set. Deleted the commented out old script for changing goals. Made it so OnTheHunt() checks if Tony is already hunting before running the rest of the function again. Idle timer now checks if Tony is idle. Hunting no longer set to false when 1.5 seconds have past after seeing player.  Goal changed to player while hunting. Speed is no longer set in the checkGoals function. CheckGoals returns true if Tony is hunting and the player is 0.5f away from him.

Vision now checks if Tony is not already hunting before calling OnTheHunt().